+++ src/lib.rs
@@ -410,8 +410,8 @@ macro_rules! fixed {
     (
         $description:expr,
         $Fixed:ident(
-            $Inner:ident, $LeEqU:tt, $UNbits:ident, $s_nbits:expr,
-            $s_nbits_m1:expr, $s_nbits_m2:expr, $s_nbits_m3:expr, $s_nbits_m4:expr
+            $Inner:ident, $LeEqU:tt, $UNbits:ident, $UNbits_m1:ident, $s_nbits:expr,
+            $s_nbits_p1:expr, $s_nbits_m1:expr, $s_nbits_m2:expr, $s_nbits_m3:expr, $s_nbits_m4:expr
         ),
         $nbytes:expr, $bytes_val:expr, $rev_bytes_val:expr, $be_bytes:expr, $le_bytes:expr,
         $IFixed:ident, $UFixed:ident, $IInner:ident, $UInner:ident, $Signedness:tt,
@@ -422,8 +422,8 @@ macro_rules! fixed {
         fixed! {
             $description,
             $Fixed[stringify!($Fixed)](
-                $Inner[stringify!($Inner)], $LeEqU, $UNbits, $s_nbits,
-                $s_nbits_m1, $s_nbits_m2, $s_nbits_m3, $s_nbits_m4
+                $Inner[stringify!($Inner)], $LeEqU, $UNbits, $UNbits_m1, $s_nbits,
+                $s_nbits_p1, $s_nbits_m1, $s_nbits_m2, $s_nbits_m3, $s_nbits_m4
             ),
             $nbytes, $bytes_val, $rev_bytes_val, $be_bytes, $le_bytes,
             $IFixed[stringify!($IFixed)], $UFixed[stringify!($UFixed)],
@@ -436,8 +436,8 @@ macro_rules! fixed {
     (
         $description:expr,
         $Fixed:ident[$s_fixed:expr](
-            $Inner:ident[$s_inner:expr], $LeEqU:tt, $UNbits:ident, $s_nbits:expr,
-            $s_nbits_m1:expr, $s_nbits_m2:expr, $s_nbits_m3:expr, $s_nbits_m4:expr
+            $Inner:ident[$s_inner:expr], $LeEqU:tt, $UNbits:ident, $UNbits_m1:ident, $s_nbits:expr,
+            $s_nbits_p1:expr, $s_nbits_m1:expr, $s_nbits_m2:expr, $s_nbits_m3:expr, $s_nbits_m4:expr
         ),
         $nbytes:expr, $bytes_val:expr, $rev_bytes_val:expr, $be_bytes:expr, $le_bytes:expr,
         $IFixed:ident[$s_ifixed:expr], $UFixed:ident[$s_ufixed:expr],
@@ -552,7 +552,10 @@ assert_eq!(two_point_75.to_string(), \"2.8\");
 
         // inherent methods that do not require Frac bounds, some of which can thus be const
         fixed_no_frac! {
-            $Fixed[$s_fixed]($Inner[$s_inner], $LeEqU, $UNbits, $s_nbits, $s_nbits_m1, $s_nbits_m2),
+            $Fixed[$s_fixed](
+                $Inner[$s_inner], $LeEqU, $UNbits, $UNbits_m1,
+                $s_nbits, $s_nbits_p1, $s_nbits_m1, $s_nbits_m2, $s_nbits_m3
+            ),
             $nbytes, $bytes_val, $rev_bytes_val, $be_bytes, $le_bytes,
             $IFixed[$s_ifixed], $UFixed[$s_ufixed],
             $IInner, $UInner, $Signedness,
@@ -574,7 +577,7 @@ assert_eq!(two_point_75.to_string(), \"2.8\");
 
 fixed! {
     "An eight",
-    FixedU8(u8, LeEqU8, U8, "8", "7", "6", "5", "4"),
+    FixedU8(u8, LeEqU8, U8, U7, "8", "9", "7", "6", "5", "4"),
     1, "0x12", "0x12", "[0x12]", "[0x12]",
     FixedI8, FixedU8, i8, u8, Unsigned,
     U8, U7, U6, U5,
@@ -582,7 +585,7 @@ fixed! {
 }
 fixed! {
     "A 16",
-    FixedU16(u16, LeEqU16, U16, "16", "15", "14", "13", "12"),
+    FixedU16(u16, LeEqU16, U16, U15, "16", "17", "15", "14", "13", "12"),
     2, "0x1234", "0x3412", "[0x12, 0x34]", "[0x34, 0x12]",
     FixedI16, FixedU16, i16, u16, Unsigned,
     U16, U15, U14, U13,
@@ -590,7 +593,7 @@ fixed! {
 }
 fixed! {
     "A 32",
-    FixedU32(u32, LeEqU32, U32, "32", "31", "30", "29", "28"),
+    FixedU32(u32, LeEqU32, U32, U31, "32", "33", "31", "30", "29", "28"),
     4, "0x1234_5678", "0x7856_3412", "[0x12, 0x34, 0x56, 0x78]", "[0x78, 0x56, 0x34, 0x12]",
     FixedI32, FixedU32, i32, u32, Unsigned,
     U32, U31, U30, U29,
@@ -598,7 +601,7 @@ fixed! {
 }
 fixed! {
     "A 64",
-    FixedU64(u64, LeEqU64, U64, "64", "63", "62", "61", "60"),
+    FixedU64(u64, LeEqU64, U64, U63, "64", "65", "63", "62", "61", "60"),
     8, "0x1234_5678_9ABC_DE0F", "0x0FDE_BC9A_7856_3412",
     "[0x12, 0x34, 0x56, 0x78, 0x9A, 0xBC, 0xDE, 0x0F]",
     "[0x0F, 0xDE, 0xBC, 0x9A, 0x78, 0x56, 0x34, 0x12]",
@@ -608,7 +611,7 @@ fixed! {
 }
 fixed! {
     "A 128",
-    FixedU128(u128, LeEqU128, U128, "128", "127", "126", "125", "124"),
+    FixedU128(u128, LeEqU128, U128, U127, "128", "129", "127", "126", "125", "124"),
     16, "0x1234_5678_9ABC_DEF0_0102_0304_0506_0708",
     "0x0807_0605_0403_0201_F0DE_BC9A_7856_3412",
     "[0x12, 0x34, 0x56, 0x78, 0x9A, 0xBC, 0xDE, 0xF0, \
@@ -621,7 +624,7 @@ fixed! {
 }
 fixed! {
     "An eight",
-    FixedI8(i8, LeEqU8, U8, "8", "7", "6", "5", "4"),
+    FixedI8(i8, LeEqU8, U8, U7, "8", "9", "7", "6", "5", "4"),
     1, "0x12", "0x12", "[0x12]", "[0x12]",
     FixedI8, FixedU8, i8, u8, Signed,
     U7, U6, U5, U4,
@@ -629,7 +632,7 @@ fixed! {
 }
 fixed! {
     "A 16",
-    FixedI16(i16, LeEqU16, U16,"16", "15", "14", "13", "12"),
+    FixedI16(i16, LeEqU16, U16, U15, "16", "17", "15", "14", "13", "12"),
     2, "0x1234", "0x3412", "[0x12, 0x34]", "[0x34, 0x12]",
     FixedI16, FixedU16, i16, u16, Signed,
     U15, U14, U13, U12,
@@ -637,7 +640,7 @@ fixed! {
 }
 fixed! {
     "A 32",
-    FixedI32(i32, LeEqU32, U32, "32", "31", "30", "29", "28"),
+    FixedI32(i32, LeEqU32, U32, U31, "32", "33", "31", "30", "29", "28"),
     4, "0x1234_5678", "0x7856_3412", "[0x12, 0x34, 0x56, 0x78]", "[0x78, 0x56, 0x34, 0x12]",
     FixedI32, FixedU32, i32, u32, Signed,
     U31, U30, U29, U28,
@@ -645,7 +648,7 @@ fixed! {
 }
 fixed! {
     "A 64",
-    FixedI64(i64, LeEqU64, U64, "64", "63", "62", "61", "60"),
+    FixedI64(i64, LeEqU64, U64, U63, "64", "65", "63", "62", "61", "60"),
     8, "0x1234_5678_9ABC_DE0F", "0x0FDE_BC9A_7856_3412",
     "[0x12, 0x34, 0x56, 0x78, 0x9A, 0xBC, 0xDE, 0x0F]",
     "[0x0F, 0xDE, 0xBC, 0x9A, 0x78, 0x56, 0x34, 0x12]",
@@ -655,7 +658,7 @@ fixed! {
 }
 fixed! {
     "A 128",
-    FixedI128(i128, LeEqU128, U128, "128", "127", "126", "125", "124"),
+    FixedI128(i128, LeEqU128, U128, U127, "128", "129", "127", "126", "125", "124"),
     16, "0x1234_5678_9ABC_DEF0_0102_0304_0506_0708",
     "0x0807_0605_0403_0201_F0DE_BC9A_7856_3412",
     "[0x12, 0x34, 0x56, 0x78, 0x9A, 0xBC, 0xDE, 0xF0, \